requirements.txt
SpeechRecognition==3.10.0
pyttsx3==2.90
python-decouple==3.8
pytest==7.4.3
pytest-mock==3.12.0
pytest-cov==4.1.0
pylint==3.0.3
coverage==7.4.1
comtypes; sys_platform == "win32"